requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
Flask-Dance==7.0.0
oauthlib==3.2.2
requests==2.31.0
gunicorn==21.2.0
serverless-wsgi==3.0.4
firebase-admin>=6.0.0
